+++ pkg/utils/models/product.go
@@ -1,14 +1,14 @@
 package models
 
 type ProductOffer struct {
-	ID              uint    `json:"id"`
-	CategoryID      int     `json:"category_id"`
-	Image           string  `json:"image"`
-	ProductName     string  `json:"product_name"`
-	Size            string  `json:"size"`
-	Stock           int     `json:"stock"`
-	Price           float64 `json:"price"`
-	DiscountedPrice float64 `json:"discounted_price"`
+	ID              uint   `json:"id"`
+	CategoryID      int    `json:"category_id"`
+	Image           string `json:"image"`
+	ProductName     string `json:"product_name"`
+	Size            string `json:"size"`
+	Stock           int    `json:"stock"`
+	Price           uint64 `json:"price"`
+	DiscountedPrice uint64 `json:"discounted_price"`
 }
 
 type Banner struct {
